Refactor my_fork and parse_args functions; update run_shell signature to include argv parameter
